Added logging feature to TSV_extract

+++ Python_Scripts/UsedForUpdate/dev/TSV_extract.py
@@ -3,7 +3,7 @@
 File: TSV_extract.py
 Author: Matt Versaggi
 Modified By: N/A
-Last Modified: January 26, 2016
+Last Modified: February 5, 2016
 
 Description:
     - Downloads dbwebexport.zip (CAL-ACCESS raw data) from the Secretary of
@@ -18,45 +18,51 @@ Sources:
     - http://campaignfinance.cdn.sos.ca.gov/dbwebexport.zip (Zip)
 '''
 
-import os
-import subprocess
-import sys
 import zipfile
+import subprocess
+import os
+from datetime import datetime
 
 zipURL = "http://campaignfinance.cdn.sos.ca.gov/dbwebexport.zip"
 tsvPath = "CalAccess/DATA/CVR_REGISTRATION_CD.TSV"
 zipName = "dbwebexport.zip"
 
 '''
-Retrieves dbwebexport.zip and extracts the .TSV file from it
+Attempts to download dbwebexport.zip and extract the .TSV file from it,
+    logging the result in TSV_extract_log.txt
+
+If successful, continues with extracting the .TSV and cleaning up after.
+If unsuccessful, removes the partial zip (if existant) and records return code
 '''
 def get_zip():
-    # Downloads the zip and places it in pwd of this script
-    wget_process = subprocess.Popen("wget -t 10 " + zipURL,
-                                    shell=True,
-                                    stderr=subprocess.PIPE)
-    _, err = wget_process.communicate()
-    if wget_process.returncode != 0:
-      #sys.stderr.write(err)
-      sys.exit(0)
-    
-    calZip = zipfile.ZipFile(zipName, 'r')
-    calZip.extract(tsvPath)
-    calZip.close()
+    # Attempts to quietly download the zip, giving up after 10 attempts
+    returnCode = subprocess.call("wget -q -t 10 " + zipURL, shell=True)
 
+    with open("TSV_extract_log.txt", 'a') as logFile:
+        if returnCode == 0:
+            logFile.write("On {0} at {1}: TSV download successful\n"
+                          .format(datetime.date(datetime.now()),
+                                  datetime.time(datetime.now())))
+            with zipfile.ZipFile(zipName, 'r') as calZip:
+                calZip.extract(tsvPath)
+            cleanup()
+        else:
+            logFile.write("On {0} at {1}: TSV download failed, returned wget error code {2}\n"
+                          .format(datetime.date(datetime.now()),
+                                  datetime.time(datetime.now()), returnCode))
+    if os.path.exists("./" + zipName):
+        subprocess.call("rm -f " + zipName, shell=True)
+        
 '''
 Pulls the .TSV out of the CalAccess directory tree and places it in the pwd,
-    removing the empty directory tree and dbwebexport.zip before returning
+    removing the empty directory tree before returning
 '''
 def cleanup():
     subprocess.call("mv " + tsvPath + " .", shell=True)
     subprocess.call("rm -r CalAccess", shell=True)
-    subprocess.call("rm -f " + zipName, shell=True)
 
 def main():
-    os.chdir('/home/data_warehouse_common/scripts')
     get_zip()
-    cleanup()
 
 if __name__ == '__main__':
     main()
